change to 30 second change

+++ wave.py
@@ -53,7 +53,7 @@ _channels = len(GPIO)
 
 _dc=[0]*_channels
 
-_micros=1000000/FREQ
+_micros=int(1000000/FREQ)
 
 old_wid = None
